+++ packages/http/axios/index.ts
@@ -1,7 +1,7 @@
 // axios配置  可自行根据项目进行更改，只需更改该文件即可，其他文件可以不动
 // The axios configuration can be changed according to the project, just change the file, other files can be left unchanged
 
-import type { AxiosRequestConfig, AxiosResponse } from 'axios';
+import type { AxiosResponse } from 'axios';
 import type { AxiosTransform, CreateAxiosOptions } from './axios-transform';
 import { VAxios } from './axios';
 import { check_status } from './check-status';
@@ -128,15 +128,9 @@ export const defaultTransform: AxiosTransform = {
         if (options.customTransform && options.customTransform.customRequest) {
             config = options.customTransform.customRequest(
                 config
-            ) as SelectPartial<
-            AxiosRequestConfig<any>,
-            'url' | 'headers' | 'method'
-            >;
+            );
         }
-        config = setTokenToHeader(options, config) as SelectPartial<
-        AxiosRequestConfig<any>,
-        'url' | 'headers' | 'method'
-        >;
+        config = setTokenToHeader(options, config) ;
         return config;
     },