+++ packages/vue3-antd-pc-ui/src/q-upload/src/types/index.ts
@@ -1,0 +1,9 @@
+
+export interface ILimit {
+    type: 'size' | 'ratio' | 'maxSize', // size为固定宽高，ratio为各比例
+    width?: number,
+    height?: number,
+    minDuration?: number,
+    maxDuration?: number,
+    message?: string
+}